Update to a1.1.1
Added error logging

+++ findTime.py
@@ -24,4 +24,4 @@ def findTime(string):
             time = 2629746 * duration
     else:
         time = 60 * int(string)
-    return float(time)
+    return int(time)
